modify oneToOne comments
